fix: move invest portfolio to profile section
